// fe/src/types/generated/SortingQuery.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Standard sorting query parameters
 */
export type SortingQuery = {
    /**
     * Field to sort by
     */
    sort_by: string | null
    /**
     * Sort order (asc or desc)
     */
    sort_order: string | null
}